refactor: reorganize component structure for scalability
